feat: add create_default_v8_encoder()
We need a slightly customised V8 serialization format encoder to handle
int and float consistently in write sum/min/max operations.
